fix render/present synchronization bugs
The Vulkan debugging validation layer was reporting two kinds of errors:

- "render finished" semaphores may not be reset, after the first few frames
- Cannot destroy unfinished "in flight" fences, at program termination

The first problem is because there is no way to explicitly wait on an "render finished"
semaphore. When the number of swap chain images (N) exceeds the number of frames in flight (M), our
first attempt to signal one occurs after M frames. Since we tell the Vulkan API towait for one of
these semaphores before presenting each swap chain image, using one semaphore for each image
guarantees every "render finished" semaphore is reset before it is reused, regardless of the number
of frames in flight.

The second problem was because we were simply not waiting for the "in flight" fences to finish
during application destruction. We solved this one easily by using the "device waitIdle" method
provided by the Vulkan API.