add transfer queue and staged vertex buffer creation
To send vertices through the graphics pipeline, we must first upload vertex data to the GPU. The
destination memory is slow because it must be made accessible to the host (CPU) and so typically
resides in the host's RAM. To speed up rendering, we can use this original vertex buffer as a
"staging ground" for transferring the vertex data into GPU RAM. This is what the transfer queue is
for.

Some devices provide a dedicated transfer queue that is separate from the graphics queue. In such
cases, we choose the dedicated transfer queue, which may improve GPU utilization by enabling
asynchronous data transfers. Otherwise, we fall back to the first transfer queue available.