begin refactoring the renderer
The Context class has become unwieldy. We need a more clear separation of Vulkan resource management
and client resource management. For example, the client should be able to build the varioous types
of buffers freely at any time before a frame is submitted for rendering without thinking about
queues, and the renderer should be able to batch queue those buffers without intimate knowledge of
their contents.

To coordinate this separation, we need dedicated resource management classes, e.g., a "descriptor
manager" for uniform buffers and a "command buffer manager" for vertex and index buffers, that
decouple resource allocation from consumption. The client can then construct whatever buffers are
needed to build the frame, and the renderer can query the management abstractions before rendering
the frame without needing the client to give that information as method arguments.

- Reorganizes Application and Main for a simpler entry point
- Drops the ApplicationBuilder class
- Extracts a "Window" class from Application that wraps the SDL_Window pointer
- Extracts a "PhysicalDevice" class from Context
- Reimplements instance, surface, and physical device constructon in Context
- Missing: device, swap chain, pipeline, and all buffer constructions from Context