add CommandPoolAllocator and CommandBufferManager classes
These classes encapsulate command pool and command buffer operations. It provides a single interface
for construction of multiple buffer types (graphics, transfer) for several usage profiles (static,
dynamic), and a convient way to reset all dynamic buffers with a single method call.

The CommandBufferManager contains many CommandPoolAllocator instances and is owned by the
Renderer. These objects are now constructed during Renderer initialization and used to create the
SwapChain, but they are not yet used in the rendering loop.