bugfix: black screen, not drawing triangles
With RenderDoc, I was able to verify that vertex data was entering the shader pipeline but no
fragments were coming out. I checked the pipeline manager's instantiation code and discovered that
it was trying to load one SPIR-V module per shader stage - in this case, two: vertex and
fragment. Since we currently use only one module per shader pipeline, it was configuring the vertex
shader but not the fragment shader. I fixed it by using the single module to configure both stages.

The client is now correctly drawing indexed vertex data with a lambda that takes a command buffer
ready for drawing.