add vertex input state generation to PipelineManager
- Moved vertex input state create info from client code into PipelineManager generator.
- Added vertex input state create info to pipeline cache map.
- Fixed a small bug in swap chain recreation.

It turns out that we can almost completely generate vertex input desccriptions. Since this data is
no longer included in the shader module configs structure, it must be generated before doing a
pipeline cache lookup. In order to include the data without polluting the client's config, we
include it as a separate component of the lookup key.