split MVP uniform buffer into separate M and VP buffers (HACK)
Still doing all the descriptor-related allocations in the client, and the work is hard-coded to
account for the two separate buffers. Next step is to replace the hard-coded allocation logic with
code that iterates over a collection of buffer configurations. Then maybe we can automate it
somewhere in the renderer.