add orstr utility function
Takes two numbers and two strings. Resturns the first string if the numbers are equal and the second
string otherwise. Since its most common use case is for adding an "s" to non-singular units (e.g.,
"byte" versus "bytes"), the second number defaults to 1 and the strings default to "" or "s".